% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCMPrior.R
\name{AddPrior}
\alias{AddPrior}
\title{Add a prior to an object}
\usage{
AddPrior(o, member = "", enclos = "?", prior, inplace = TRUE)
}
\arguments{
\item{o}{an object.}

\item{member}{a member expression. Member expressions are character strings
denoting named elements in a list object (see examples). Default: "".}

\item{enclos}{a character string containing the special symbol '?'. This
symbol is to be replaced by matching expressions. The result of this
substitution can be anything but, usually would be a valid R expression.
Default: "?".}

\item{prior}{a prior object.}

\item{inplace}{logical (TRUE by default) indicating if the attribute should
be set to the object in the current environment, or a modified object should
be returned.}
}
\value{
if inplace is TRUE (default) nothing is returned. Otherwise, a
modified version of \code{o} is returned.
}
\description{
This is an S3 generic with methods depending on the S3 class of the \code{o},
argument.
}
\examples{
library(PCMBase)
model <- PCMBaseTestObjects$model_MixedGaussian_ab

AddPrior(
  model,
  prior = ParameterPrior(
    d = "dunif", r = "runif",
    p = list(min = PCMParamGetShortVector(PCMParamLowerLimit(model)),
             max = PCMParamGetShortVector(PCMParamUpperLimit(model)))))

AddPrior(
  model, "a$Sigma_x", enclos = "diag(?[,,1])",
  prior = ParameterPrior(
    d = "dunif", r = "runif",
    p = list(min = c(0.2, 0.1, 0.1), max = c(5, 5, 5))))

modelPrior <- PCMPrior(model)
vec <- PCMParamGetShortVector(model)
PriorDensity(modelPrior, vec)
}
