% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCState.R
\name{as.MCMCState}
\alias{as.MCMCState}
\title{If necessary, convert an object to an MCMCState object.}
\usage{
as.MCMCState(s, ctx = NULL)
}
\arguments{
\item{s}{a \code{\link{MCMCState}} or a \code{\link{MCMCStateVector}}
object.}

\item{ctx}{a \code{\link{MCMCContext}} object. This is needed only if
\code{s} is \code{\link{MCMCStateVector}} object.}
}
\value{
an MCMCState object corresponding to \code{s}. If \code{s} is already
a \code{\link{MCMCState}} object, it is returned as is. Otherwise, a
\code{\link{MCMCState}} object is constructed from \code{s} and \code{ctx}.
}
\description{
If necessary, convert an object to an MCMCState object.
}
