% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGPMState.R
\name{as.MGPMState}
\alias{as.MGPMState}
\title{If necessary, convert an object to an MGPMState object.}
\usage{
as.MGPMState(s, ctx = NULL)
}
\arguments{
\item{s}{a \code{\link{MGPMState}} or a \code{\link{MGPMStateVector}}
object.}

\item{ctx}{a \code{\link{MGPMContext}} object. This is needed only if
\code{s} is \code{\link{MGPMStateVector}} object.}
}
\value{
an MGPMState object corresponding to \code{s}. If \code{s} is already
a \code{\link{MGPMState}} object, it is returned as is. Otherwise, a
\code{\link{MGPMState}} object is constructed from \code{s} and \code{ctx}.
}
\description{
If necessary, convert an object to an MGPMState object.
}
