% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCState.R
\name{is.MCMCState}
\alias{is.MCMCState}
\title{Check if an object is of S3 class 'MCMCState'}
\usage{
is.MCMCState(o)
}
\arguments{
\item{o}{an object.}
}
\value{
a logical.
}
\description{
Check if an object is of S3 class 'MCMCState'
}
