% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCState.R
\name{is.MCMCStateVector}
\alias{is.MCMCStateVector}
\title{Check if an object is of S3 class 'MCMCStateVector'}
\usage{
is.MCMCStateVector(o)
}
\arguments{
\item{o}{an object.}
}
\value{
a logical.
}
\description{
Check if an object is of S3 class 'MCMCStateVector'
}
