% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGPMState.R
\name{is.MGPMState}
\alias{is.MGPMState}
\title{Check if an object is of S3 class 'MGPMState'}
\usage{
is.MGPMState(o)
}
\arguments{
\item{o}{an object.}
}
\value{
a logical.
}
\description{
Check if an object is of S3 class 'MGPMState'
}
