% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGPMStatePrior.R
\name{is.MGPMStatePrior}
\alias{is.MGPMStatePrior}
\title{Check if an object is a MGPMStatePrior object.}
\usage{
is.MGPMStatePrior(o)
}
\arguments{
\item{o}{an R object.}
}
\value{
a logical indicating whether the S3 class of \code{o} inherits from
\code{'MGPMStatePrior'}.
}
\description{
Check if an object is a MGPMStatePrior object.
}
