% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGPMState.R
\name{is.MGPMStateVector}
\alias{is.MGPMStateVector}
\title{Check if an object is of S3 class 'MGPMStateVector'}
\usage{
is.MGPMStateVector(o)
}
\arguments{
\item{o}{an object.}
}
\value{
a logical.
}
\description{
Check if an object is of S3 class 'MGPMStateVector'
}
