% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCMPrior.R
\name{is.ParameterPrior}
\alias{is.ParameterPrior}
\title{Check if an object is a ParameterPrior object.}
\usage{
is.ParameterPrior(o)
}
\arguments{
\item{o}{an R object.}
}
\value{
a logical indicating whether the S3 class of \code{o} inherits from
\code{'ParameterPrior'}.
}
\description{
Check if an object is a ParameterPrior object.
}
