% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCMPrior.R
\name{is.PCMPrior}
\alias{is.PCMPrior}
\title{Check if an object is a PCMPrior object.}
\usage{
is.PCMPrior(o)
}
\arguments{
\item{o}{an R object.}
}
\value{
a logical indicating whether the S3 class of \code{o} inherits from
\code{'PCMPrior'}.
}
\description{
Check if an object is a PCMPrior object.
}
