% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prior.R
\name{is.Prior}
\alias{is.Prior}
\title{Check if an object is a Prior object.}
\usage{
is.Prior(o)
}
\arguments{
\item{o}{an R object.}
}
\value{
a logical indicating whether the S3 class of \code{o} inherits from
\code{'Prior'}.
}
\description{
Check if an object is a Prior object.
}
