% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCState.R
\name{MCMCState}
\alias{MCMCState}
\title{Construct a MCMCState object from a state vector}
\usage{
MCMCState(s, ctx)
}
\arguments{
\item{s}{a numerical vector representing a MCMC state. Some of the
entries in this vector are interpreted as integers denoting nodes in a
tree, regimes or model types (see \code{\link{MCMCStateVector}}).}

\item{ctx}{a MCMCContext object.}
}
\value{
a MCMCState object.
}
\description{
Construct a MCMCState object from a state vector
}
\seealso{
\code{\link{MCMCStateVector}}
}
