% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGPMState.R
\name{MGPMState}
\alias{MGPMState}
\title{Construct a MGPMState object from a state vector}
\usage{
MGPMState(s, ctx)
}
\arguments{
\item{s}{a numerical vector representing a MGPM state. Some of the
entries in this vector are interpreted as integers denoting nodes in a
tree, regimes or model types (see \code{\link{MGPMStateVector}}).}

\item{ctx}{a MGPMContext object.}
}
\value{
a MGPMState object.
}
\description{
Construct a MGPMState object from a state vector
}
\seealso{
\code{\link{MGPMStateVector}}
}
