% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGPMStatePrior.R
\name{MGPMStatePrior}
\alias{MGPMStatePrior}
\title{Create a MGPMStatePrior object for a MGPMState object}
\usage{
MGPMStatePrior(state, ...)
}
\arguments{
\item{state}{a MGPMState object.}

\item{...}{additional arguments to be stored as member elements in the
returned MGPMStatePrior object.}
}
\value{
an object of S3 class \code{c("MGPMStatePrior", "Prior")}.
}
\description{
Create a MGPMStatePrior object for a MGPMState object
}
