% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGPMState.R
\name{MGPMStateVector}
\alias{MGPMStateVector}
\title{Construct a MGPM state vector.}
\usage{
MGPMStateVector(K, n, l, r, m, v)
}
\arguments{
\item{K}{an integer (see Details).}

\item{n}{an integer vector of length \code{K} denoting shift nodes (see Details).}

\item{l}{a real vector of length \code{K} (see Details).}

\item{r}{an integer vector of length \code{K} (see Details).}

\item{m}{an integer vector of length \code{R}, where \code{R} denotes the
total number of regimes (see Details).}

\item{v}{a real vector of length \code{P}, where \code{P} denotes the total
number of numeric parameters of the MGPM model represented by the state
(see Details).}
}
\value{
a numerical vector representing the concatenation of
\code{K, n, l, r, m, v}, with S3 class set to 'MGPMStateVector'.
}
\description{
Construct a MGPM state vector.
}
\details{
The MGPMState of an MCMC is represented by a numerical vector:
\deqn{\vec{s}=(K, n_2,...,n_{K+1}, l_2,...,l_{K+1}, r_2,...,r_{K+1}, m_1,...,m_R, v_1,...,v_P)^T}
The element of this vector are described as follows:
\describe{
\item{\eqn{K}: }{number of shifts;}
\item{\eqn{(n_2,...,n_{K+1})^T}: }{shift nodes - the shifts in the model occur
at points within the branches leading to the shift nodes in tip-ward
direction. The corresponding locations of these points are specified by
\eqn{(l_2,...,l_{K+1})^T}. The nodes \eqn{(n_2,...,n_{K+1})^T} should be ordered
according to \code{\link{PCMTreePreorder}(ctx$tree)}.}
\item{\eqn{(l_2,...,l_{K+1})^T}: }{offset of the shift points measured as
distances from the beginnings of the branches leading to shift nodes in
tip-ward direction.}
\item{\eqn{(r_2,...,r_{K+1})^T}: }{regime index vector. This is an integer
vector with elements among \eqn{(N+1,n_2,...,n_{K+1})^T}, indicating the regime
associated with each part in the tree. The regimes are named as the shift
nodes, with N+1 corresponding to the part (and regime) originating at the
root. It is possible to have lumped regimes, that is, different parts
of the tree having the same regime. This regime-lumping must obey the
following rules:
\enumerate{
\item neighbor parts cannot have a lumped regime. Two parts originating at
nodes \eqn{n_i} and \eqn{n_j} in the tree are called neighbor parts if they
are separated solely by \eqn{n_i} or by \eqn{n_j};
\item to resolve the conflict between the shift nodes of the different parts
covered by a lumped regime, it is established that the name of a lumped
regime must equal the shift-node that appears first, according to
\code{\link{PCMTreePreorder}(ctx$tree)}.
}}
\item{\eqn{(m_1,...,m_R)^T}: }{model type assignment to the unique regimes.
This is an integer vector with elements between 1 and M, M denoting the
total number of model types possible. Each element corresponds to an
element in \code{unique(c(N+1,r_2,...,r_{K+1}))}}
\item{\eqn{(v_1,...,v_P)^T}: }{real numbers passed to
\code{\link{PCMParamLoadOrStore}}. This is a vectorized form of all model
parameters.}
}
}
\seealso{
\code{\link{MGPMState}}
}
