% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParameterPrior.R
\name{ParameterPrior}
\alias{ParameterPrior}
\title{Create a ParameterPrior object}
\usage{
ParameterPrior(d = NULL, r = NULL, p = NULL, p.d = NULL,
  p.r = NULL, ...)
}
\arguments{
\item{d, r}{character strings denoting the names of a density function and of
a random generator function for some distribution, e.g. "dunif" and "runif".
Alternatively, these can be function objects, but this is not recommended,
because it can result in multiple duplication of these objects in memory.
Default values: \code{NULL}.}

\item{p}{a list with 'static' arguments common for \code{d} and \code{r}.
By 'static', it is meant arguments that do not change during a sampling
procedure, e.g. the rate of an exponentional prior or the boundaries of a
uniform prior. Default value: \code{NULL}}

\item{p.d}{a list of static arguments to be passed specifically to
\code{d} (added to \code{p}). Default value: \code{NULL}.}

\item{p.r}{a list of static arguments to be passed specifically to
\code{r} (added to \code{p}). Default value: \code{NULL}.}

\item{...}{additional arguments to be stored as member elements in the
returned ParameterPrior object.}
}
\value{
an object of S3 class \code{c("ParameterPrior", "Prior")}.
}
\description{
Create a ParameterPrior object
}
