% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCMPrior.R
\name{PCMPrior}
\alias{PCMPrior}
\title{Create a PCMPrior object for a PCM model object}
\usage{
PCMPrior(model, ...)
}
\arguments{
\item{model}{a PCM object.}

\item{...}{additional arguments to be stored as member elements in the
returned PCMPrior object.}
}
\value{
an object of S3 class \code{c("PCMPrior", "CompositePrior", "Prior")}.
}
\description{
Create a PCMPrior object for a PCM model object
}
