% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGPMState.R
\name{MGPMState}
\alias{MGPMState}
\title{(Re-)construct an MGPMState object from a state vector}
\usage{
MGPMState(s, ctx)
}
\arguments{
\item{s}{a numerical vector representing an MGPM state. Some of the entries
in this vector are interpreted as integers denoting nodes in a tree,
regimes or model types (see Details).}

\item{ctx}{a MGPMContext object.}
}
\value{
a MGPMState object.
}
\description{
(Re-)construct an MGPMState object from a state vector
}
\seealso{
\code{\link{MGPMStateVector}}
}
