% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCMPrior.R
\name{ParameterPrior}
\alias{ParameterPrior}
\title{Create a ParameterPrior object}
\usage{
ParameterPrior(enclos = "?", pos = NULL, d = NULL, r = NULL,
  p = NULL, p.d = NULL, p.r = NULL, ...)
}
\arguments{
\item{enclos}{a character string containing the special symbol '?'. This
symbol is to be replaced by expressions denoting objects to which this
ParameterPrior object will be added. The result of this
substitution can be anything but, usually would be a valid R expression.
Default: "?". See also \code{\link{AddPrior}}.}

\item{pos}{an integer denoting an offset position in a parameter vector.
Default: \code{NULL}.}

\item{d, r}{character strings denoting the names of a density function and of
a random generator function for some distribution, e.g. "dunif" and "runif".}

\item{p}{a list with 'static' arguments common for \code{d} and \code{r}.
By 'static', it is meant arguments that do not change during a sampling
procedure, e.g. the rate of an exponentional prior or the boundaries of a
uniform prior.}

\item{p.d}{a list of static arguments to be passed specifically to
\code{d} (added to \code{p}).}

\item{p.r}{a list of static arguments to be passed specifically to
\code{r} (added to \code{p}).}

\item{...}{additional arguments to be stored as member elements in the
returned ParameterPrior object.}
}
\value{
an object of S3 class \code{c("ParameterPrior", "Prior")}.
}
\description{
Create a ParameterPrior object
}
