% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCState.R
\name{posK}
\alias{posK}
\alias{posn}
\alias{posl}
\alias{posr}
\alias{posm}
\alias{posv}
\title{Indices of different parts of an MCMCState in an MCMCStateVector}
\usage{
posK(s, ctx = NULL)

posn(s, ctx = NULL)

posl(s, ctx = NULL)

posr(s, ctx = NULL)

posm(s, ctx = NULL)

posv(s, ctx = NULL)
}
\arguments{
\item{s}{a \code{\link{MCMCState}} or a \code{\link{MCMCStateVector}}
object.}

\item{ctx}{a \code{\link{MCMCContext}} object. This is needed only if
\code{s} is \code{\link{MCMCStateVector}} object.}
}
\value{
an integer vector.
}
\description{
The different parts of an MCMCState are described in
\code{\link{MCMCStateVector}}. \code{posK} returns the osition of the
 number of shifts (this is always equal to 1). See the section functions
 for the others.
}
\section{Functions}{
\itemize{
\item \code{posn}: Positions of the shift nodes;

\item \code{posl}: Positions of the offsets of the shift-points in tip-ward direction
relative to the beginnings of the branches leading to shift nodes;

\item \code{posr}: Position of the regime id's for each shift node;

\item \code{posm}: Positions of the model type id's for each regime;

\item \code{posv}: Positions of the model parameters.
}}

\seealso{
\code{\link{MCMCState}}, \code{\link{MCMCStateVector}}
}
